pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ZeroTwo-M1000 <BioArtex@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
prisma = "^0.12.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
